fix(locate): accept float rssi (앱 평균값 대응)
앱이 '최근 N개 평균'을 보내면 rssi 가 float 가 되는데, 기존 검증이
int 만 허용해 INVALID_PAYLOAD 로 거부됐음. int·float 모두 수용하도록
변경 (bool 은 명시적 배제). WifiSample.rssi: int → float.

- subway_server/api/locate.py: 검증 완화 + float() 정규화
- subway_server/core/locator.py: WifiSample.rssi 타입 widening
- tests: float 허용 / bool 거부 케이스 2건 추가 (61 passed)
- docs/05-API명세.md §5.2.3: rssi 타입 number 로 정정

통합 경계 contract(estimate 시그니처)는 불변. 화경 KNN 영향 없음.

+++ subway_server/core/locator.py
@@ -11,7 +11,7 @@ from typing import Callable
 @dataclass(frozen=True)
 class WifiSample:
     bssid: str
-    rssi: int
+    rssi: float  # int·float 모두 허용 (앱이 최근 N개 평균을 보내면 float)
 
 
 LocationEstimator = Callable[[list[WifiSample]], str]
